drivers/net: Add Aspeed nic driver
This is an alternate driver for the ftgmac100 hardware found in the
Aspeed BMCs. It contains support for NC-SI that is lacking in the
existing ftgmac100 driver.

The code comes from the u-boot tree used by OpenBMC, which has origins
in the Facebook OpenBMC project.

In addition the following fixes were made:
 - Port new release
 - Clean up hw init
 - Remove scratch register writes
 - Fix RX buffer size
 - Make rx fifo poll for data (was: Fix unreliable networking)
 - Create aspeed_write_hwaddr from set_mac_address
 - Do not start hardware in initialize
 - Do not fill in a random MAC address
 - Start mac after ring buffers are initialized

Signed-off-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Cédric Le Goater <clg@kaod.org>